fix: eslint searchbar complexity error fixed by seperating some elements
