+++ src/modules/home/components/header/constants/languages.constant.ts
@@ -1,5 +1,5 @@
 export const languages = [
-    { code: 'uz', name: "O'zbek", region: "O'zbekiston" },
-    { code: 'en', name: 'English', region: 'United States' },
-    { code: 'ru', name: 'Русский', region: 'Россия' },
+  { code: 'uz', name: "O'zbek", region: "O'zbekiston" },
+  { code: 'en', name: 'English', region: 'United States' },
+  { code: 'ru', name: 'Русский', region: 'Россия' },
 ];
