+++ src/modules/home/components/header/search-bar.tsx
@@ -2,32 +2,12 @@ import { useState } from 'react';
 import Box from '@mui/material/Box';
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
-import Popover from '@mui/material/Popover';
-import TextField from '@mui/material/TextField';
-import List from '@mui/material/List';
-import ListItem from '@mui/material/ListItem';
-import ListItemIcon from '@mui/material/ListItemIcon';
-import ListItemText from '@mui/material/ListItemText';
 import { SearchbarContainer, SearchButton, SearchSection, StyledDivider } from '../../styles';
 import SearchIcon from '@mui/icons-material/Search';
-import LocationIcon from '@mui/icons-material/LocationOn';
 import { CalendarModal } from './modals/calendar.modal';
-import { type GuestCounts, GuestsModal } from './modals/guests.modal';
+import { GuestsModal } from './modals/guests.modal';
 import { searchbarStyles } from './styles';
-
-interface DestinationInterface {
-  id: number;
-  name: string;
-  country: string;
-}
-const destinations: DestinationInterface[] = [
-  { id: 6, name: 'Cozy', country: 'Uz' },
-  { id: 1, name: 'New York', country: 'United States' },
-  { id: 2, name: 'London', country: 'United Kingdom' },
-  { id: 3, name: 'Paris', country: 'France' },
-  { id: 4, name: 'Tokyo', country: 'Japan' },
-  { id: 5, name: 'Sydney', country: 'Australia' },
-];
+import { DestinationSearch } from './destination-search';
 
 const months = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
 
@@ -36,16 +16,42 @@ interface SearchBarProps {
   setSelectedLocation: (newState: string) => void;
 }
 
-export function SearchBar({ activeNav, setSelectedLocation }: SearchBarProps) {
-  const [anchorEl, setAnchorEl] = useState<HTMLElement | null>(null);
-  const [searchValue, setSearchValue] = useState('');
-  const [selectedDestination, setSelectedDestination] = useState<{ name: string } | null>(null);
+interface DateButtonProps {
+  label: string;
+  date: Date | null;
+  onOpen: () => void;
+}
 
-  const [isCalendarOpen, setIsCalendarOpen] = useState<boolean>(false);
+function DateButton({ label, date, onOpen }: DateButtonProps) {
+  return (
+    <SearchSection sx={searchbarStyles.datesSection.commonDisplay}>
+      <Stack alignItems="flex-start" sx={searchbarStyles.datesSection.commonWidth} onClick={onOpen}>
+        <Typography sx={searchbarStyles.commonTypography.title}>{label}</Typography>
+        <Typography sx={searchbarStyles.commonTypography.subtitle}>
+          {date ? `${months[date.getMonth()]} ${date.getDate()}` : `Add ${label.toLowerCase()}`}
+        </Typography>
+      </Stack>
+    </SearchSection>
+  );
+}
+
+export function SearchBar({ activeNav, setSelectedLocation }: SearchBarProps) {
+  const [searchValue, setSearchValue] = useState('');
   const [startDate, setStartDate] = useState<Date | null>(null);
   const [endDate, setEndDate] = useState<Date | null>(null);
+  const [isCalendarOpen, setIsCalendarOpen] = useState<boolean>(false);
   const [isGuestsModalOpen, setIsGuestsModalOpen] = useState(false);
   const [guestCounts, setGuestCounts] = useState({ adults: 1, children: 0, infants: 0, pets: 0 });
+
+  const formatGuestCount = () => {
+    const total = guestCounts.adults + guestCounts.children;
+    const parts = [`${total} guest${total !== 1 ? 's' : ''}`];
+
+    if (guestCounts.infants > 0) parts.push(`${guestCounts.infants} infant${guestCounts.infants !== 1 ? 's' : ''}`);
+    if (guestCounts.pets > 0) parts.push(`${guestCounts.pets} pet${guestCounts.pets !== 1 ? 's' : ''}`);
+
+    return parts.join(', ');
+  };
 
   const handleDateSelect = (start: Date | null, end: Date | null) => {
     setStartDate(start);
@@ -53,55 +59,21 @@ export function SearchBar({ activeNav, setSelectedLocation }: SearchBarProps) {
     setIsCalendarOpen(false);
   };
 
-  const handleSearchClick = () => {
-    const search = searchValue.trim() || selectedDestination?.name || '';
-    if (!search) {
-      setSearchValue('');
-      setSelectedDestination(null);
+  const renderDateSection = () => {
+    if (activeNav === 'stays') {
+      return (
+        <Stack
+          direction="row"
+          spacing={searchbarStyles.datesSection.checkInOutStack.spacing}
+          divider={<StyledDivider orientation="vertical" flexItem={true} variant="middle" />}
+          sx={searchbarStyles.datesSection.checkInOutStack.styles}
+        >
+          <DateButton label="Check in" date={startDate} onOpen={() => setIsCalendarOpen(true)} />
+          <DateButton label="Check out" date={endDate} onOpen={() => setIsCalendarOpen(true)} />
+        </Stack>
+      );
     }
-    setSearchValue(search);
-    setSelectedLocation(search);
-  };
-
-  const handleClose = () => {
-    if (!searchValue.trim()) {
-      setSelectedDestination(null);
-      setSearchValue('');
-    }
-    setAnchorEl(null);
-  };
-
-  const handleDestinationSelect = (destination: DestinationInterface) => {
-    if (destination.name) setSearchValue(destination.name);
-    setSelectedDestination(destination);
-    setSearchValue(destination.name);
-    setSelectedLocation(destination.name);
-    handleClose();
-  };
-
-  const open = Boolean(anchorEl);
-
-  const formatGuestCount = () => {
-    const total = guestCounts.adults + guestCounts.children;
-    let text = `${total} guest${total !== 1 ? 's' : ''}`;
-
-    if (guestCounts.infants > 0) {
-      text += `, ${guestCounts.infants} infant${guestCounts.infants !== 1 ? 's' : ''}`;
-    }
-    if (guestCounts.pets > 0) {
-      text += `, ${guestCounts.pets} pet${guestCounts.pets !== 1 ? 's' : ''}`;
-    }
-    return text;
-  };
-
-  const filteredDestinations = destinations.filter(
-    (dest: DestinationInterface) =>
-      dest.name.toLowerCase().includes(searchValue.toLowerCase()) ||
-      dest.country.toLowerCase().includes(searchValue.toLowerCase())
-  );
-
-  const handleGuestCountsChange = (counts: GuestCounts) => {
-    setGuestCounts(counts);
+    return <DateButton label="Date" date={startDate} onOpen={() => setIsCalendarOpen(true)} />;
   };
 
   return (
@@ -112,111 +84,9 @@ export function SearchBar({ activeNav, setSelectedLocation }: SearchBarProps) {
         divider={<StyledDivider orientation="vertical" flexItem={true} variant="middle" />}
         sx={searchbarStyles.stack.styles}
       >
-        <SearchSection onClick={(e) => setAnchorEl(e.currentTarget)} sx={searchbarStyles.searchSection.container}>
-          <Stack alignItems="flex-start" sx={searchbarStyles.searchButtonContainer}>
-            <Typography sx={searchbarStyles.commonTypography.title}>Search here</Typography>
-            <Typography sx={searchbarStyles.commonTypography.subtitle}>
-              {searchValue || (selectedDestination?.name ?? 'Search destinations')}
-            </Typography>
-          </Stack>
-          <Popover
-            open={open}
-            anchorEl={anchorEl}
-            onClose={handleClose}
-            anchorOrigin={searchbarStyles.searchSection.popover.anchorOrigin}
-            transformOrigin={searchbarStyles.searchSection.popover.transformOrigin}
-            PaperProps={{
-              sx: searchbarStyles.searchSection.popover.paperProps,
-            }}
-          >
-            <Box sx={{ p: 2 }}>
-              <TextField
-                autoFocus={true}
-                fullWidth={true}
-                value={searchValue}
-                onChange={(e) => setSearchValue(e.target.value)}
-                placeholder="Search destinations"
-                variant="outlined"
-                size="small"
-                sx={searchbarStyles.searchSection.popover.searchDestionationsTextField}
-              />
-            </Box>
-            <List sx={searchbarStyles.searchSection.popover.list}>
-              {filteredDestinations.length === 0 ? (
-                <ListItem>
-                  <ListItemText primary="No destinations found" />
-                </ListItem>
-              ) : (
-                filteredDestinations.map((destination) => (
-                  <ListItem
-                    key={destination.id}
-                    onClick={() => handleDestinationSelect(destination)}
-                    sx={searchbarStyles.searchSection.popover.listItem}
-                  >
-                    <ListItemIcon>
-                      <LocationIcon color="action" />
-                    </ListItemIcon>
-                    <ListItemText
-                      primary={destination.name}
-                      secondary={destination.country}
-                      primaryTypographyProps={searchbarStyles.searchSection.popover.listItemText.primary}
-                      secondaryTypographyProps={searchbarStyles.searchSection.popover.listItemText.secondary}
-                    />
-                  </ListItem>
-                ))
-              )}
-            </List>
-          </Popover>
-        </SearchSection>
+        <DestinationSearch searchValue={searchValue} setSearchValue={setSearchValue} setSelectedLocation={setSelectedLocation} />
 
-        <Box sx={searchbarStyles.datesSection.container}>
-          {activeNav === 'stays' ? (
-            <Stack
-              direction="row"
-              spacing={searchbarStyles.datesSection.checkInOutStack.spacing}
-              divider={<StyledDivider orientation="vertical" flexItem={true} variant="middle" />}
-              sx={searchbarStyles.datesSection.checkInOutStack.styles}
-            >
-              <SearchSection sx={searchbarStyles.datesSection.commonDisplay}>
-                <Stack
-                  alignItems="flex-start"
-                  sx={searchbarStyles.datesSection.commonWidth}
-                  onClick={() => setIsCalendarOpen(true)}
-                >
-                  <Typography sx={searchbarStyles.commonTypography.title}>Check in</Typography>
-                  <Typography sx={searchbarStyles.commonTypography.subtitle}>
-                    {startDate ? `${months[startDate.getMonth()]} ${startDate.getDate()}` : 'Add dates'}{' '}
-                  </Typography>
-                </Stack>
-              </SearchSection>
-              <SearchSection sx={searchbarStyles.datesSection.commonDisplay}>
-                <Stack
-                  alignItems="flex-start"
-                  sx={searchbarStyles.datesSection.commonWidth}
-                  onClick={() => setIsCalendarOpen(true)}
-                >
-                  <Typography sx={searchbarStyles.commonTypography.title}>Check out</Typography>
-                  <Typography sx={searchbarStyles.commonTypography.subtitle}>
-                    {endDate ? `${months[endDate.getMonth()]} ${endDate.getDate()}` : 'Add dates'}
-                  </Typography>
-                </Stack>
-              </SearchSection>
-            </Stack>
-          ) : (
-            <SearchSection sx={searchbarStyles.datesSection.commonDisplay}>
-              <Stack
-                alignItems="flex-start"
-                sx={searchbarStyles.datesSection.commonWidth}
-                onClick={() => setIsCalendarOpen(true)}
-              >
-                <Typography sx={searchbarStyles.commonTypography.title}>Date</Typography>
-                <Typography sx={searchbarStyles.commonTypography.subtitle}>
-                  {startDate ? `${months[startDate.getMonth()]} ${startDate.getDate()}` : 'Select Date'}
-                </Typography>
-              </Stack>
-            </SearchSection>
-          )}
-        </Box>
+        <Box sx={searchbarStyles.datesSection.container}>{renderDateSection()}</Box>
 
         <SearchSection sx={searchbarStyles.datesSection.commonDisplay}>
           <Stack alignItems="flex-start" sx={searchbarStyles.datesSection.commonWidth} onClick={() => setIsGuestsModalOpen(true)}>
@@ -225,10 +95,11 @@ export function SearchBar({ activeNav, setSelectedLocation }: SearchBarProps) {
           </Stack>
         </SearchSection>
 
-        <SearchButton onClick={handleSearchClick} sx={searchbarStyles.searchButton}>
+        <SearchButton onClick={() => setSelectedLocation(searchValue)} sx={searchbarStyles.searchButton}>
           <SearchIcon sx={searchbarStyles.searchIcon} />
         </SearchButton>
       </Stack>
+
       <CalendarModal
         open={isCalendarOpen}
         onClose={() => setIsCalendarOpen(false)}
@@ -240,7 +111,7 @@ export function SearchBar({ activeNav, setSelectedLocation }: SearchBarProps) {
         open={isGuestsModalOpen}
         onClose={() => setIsGuestsModalOpen(false)}
         guestCounts={guestCounts}
-        onGuestCountsChange={handleGuestCountsChange}
+        onGuestCountsChange={setGuestCounts}
       />
     </SearchbarContainer>
   );
