+++ src/modules/home/components/header/styles/index.ts
@@ -1,3 +1,4 @@
-export * from "./guest-modal.styles";
-export * from "./language-modal.styles";
-export * from "./user-navigation.styles";
+export * from './guest-modal.styles';
+export * from './language-modal.styles';
+export * from './user-navigation.styles';
+export * from './search-bar.styles';
